+++ cdm_reader_mapper/mdf_reader/utils/decoders.py
@@ -90,14 +90,10 @@ class df_decoders:
         """DOCUMENTATION."""
         # Caution: int(str(np.nan),36) ==> 30191
         data = data.replace(" ", np.nan, regex=True)
-        base10 = [str(int(str(i), 36)) if i == i and i else np.nan for i in data]
-        # base10 = []
-        # for i in data:
-        #  if i == i and i:
-        #    base10 += [str(int(str(i), 36))]
-        #  else:
-        #    base10 += np.nan
-        return pd.Series(base10, dtype=self.dtype)
+
+        data = [str(int(str(i), 36)) if i == i and i else np.nan for i in data]
+
+        return pd.Series(data, dtype=self.dtype)
 
 
 decoders = dict()