+++ CHANGES.rst
@@ -15,16 +15,18 @@ Breaking changes
 * remove unused modules from ``metmetpy``
 * ``cdm_reader_mapper.mdf_reader`` split data_models into code_tables and schema
 * logging: Allow for use of log file (:pull:`6`)
+* cannot use as command-line tool anymore (:pull:`22`)
 
 Internal changes
 ^^^^^^^^^^^^^^^^
-* adding tests to cdm_reader_mapper testing suite (:pull:`2`)
+* adding tests to cdm_reader_mapper testing suite (:issue:`12`, :pull:`2`, :pull:`20`, :pull:`22`)
 * adding testing result data (:pull:`4`)
 * use slugify insted of unidecde for licening reasons
 * remove pip install instruction (:pull:`2`)
 * ``HISTORY.rst`` has been renamed ``CHANGES.rst``, to follow `xclim`-like conventions (:pull:`7`).
 * speed up mapping functions with `swifter` (:pull:`4`)
 * ``mdf_reader``: adding auxiliary functions and classes (:pull:`4`)
+* ``mdf_reader``: read tables line-by-line (:pull:`20`)
 
 Bug fixes
 ^^^^^^^^^